+++ CHANGELOG.md
@@ -3,11 +3,11 @@
 - JDK version upgrade to 1.8 (was 1.7).
 
 ### Added
-- JUnit5 extension class equivalents for all Beeju Rules. 
+- JUnit5 extension class equivalents for all BeeJU Rules. 
 
 ## [2.0.0] - 2019-08-30
 ### Added
-- Support for JUnit5 using `migration-support` dependency. NOTE - the transitive dependency for JUnit from Beeju has been removed so you must depend on it in your own POM. 
+- Support for JUnit5 using `migration-support` dependency. NOTE - the transitive dependency for JUnit from BeeJU has been removed so you must depend on it in your own POM. 
 
 ## [1.3.2] - 2019-07-10
 ### Changed